# apps/dashboard/lib/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Carbon API",
    "description": "Control plane for Carbon — projects, ingestion, emulators, snapshots, artifacts, org/billing, and CLI auth. Every endpoint accepts either an API key (`x-carbon-key`) or a Better Auth session cookie/bearer token; the docs and `/health` endpoints are the only routes served anonymously.",
    "version": "dev",
    "contact": {
      "name": "Carbon",
      "url": "https://github.com/TaxCollector23/carbon",
      "email": "support@example.com"
    },
    "license": { "name": "Apache-2.0", "url": "https://www.apache.org/licenses/LICENSE-2.0" }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-carbon-key",
        "description": "A key minted via `POST /v1/api-keys` or the `bootstrap` script. Prefix `ck_live_*` for production, `ck_test_*` for test envs."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "better-auth-session",
        "description": "A Better Auth session bearer token. Preferred for programmatic access from a signed-in user."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "The Better Auth session cookie set by the dashboard. Used implicitly when the browser calls the API cross-origin with `credentials: \"include\"`."
      }
    },
...
  },
  "paths": {
...
    "/v1/chaos-presets": {
      "get": {
        "summary": "List chaos presets",
        "tags": ["Chaos Presets"],
        "description": "Return every chaos preset visible to the caller's org, sorted by name.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "orgId": { "type": "string" },
                          "name": { "type": "string" },
                          "description": { "type": "string", "nullable": true },
                          "rules": { "type": "array", "items": {} },
                          "builtIn": { "type": "boolean" },
                          "createdAt": {
                            "anyOf": [
                              { "type": "string" },
                              { "type": "string", "format": "date-time" }
                            ]
                          }
                        },
                        "required": ["id", "orgId", "name", "rules"],
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": ["data"],
                  "additionalProperties": false
                },
                "example": {
                  "data": [
                    {
                      "id": "chaos_01HXK5H7Q9C0R3Q1S8V6M4WJZK",
                      "orgId": "org_01HXK5H7Q9C0R3Q1S8V6M4WJZK",
                      "name": "flaky-checkout",
                      "description": "10% 500s and 200ms latency on POST /orders",
                      "rules": [
                        {
                          "kind": "error",
                          "match": { "method": "POST", "path": "/orders" },
                          "probability": 0.1,
                          "status": 500
                        },
                        {
                          "kind": "latency",
                          "match": { "path": "/orders" },
                          "floorMs": 200,
                          "jitterMs": 50
                        }
                      ],
                      "builtIn": false,
                      "createdAt": "2025-11-14T18:22:41.000Z"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a chaos preset",
        "tags": ["Chaos Presets"],
        "description": "Register a named chaos preset scoped to the caller's org. Names are unique per org; duplicates return 409.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9][a-z0-9-_]*$"
                  },
                  "description": { "type": "string", "maxLength": 500 },
                  "rules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "kind": { "type": "string", "enum": ["error", "latency"] },
                        "match": {
                          "type": "object",
                          "properties": {
                            "method": { "type": "string", "minLength": 1, "maxLength": 10 },
                            "path": { "type": "string", "minLength": 1, "maxLength": 200 }
                          },
                          "additionalProperties": false
                        },
                        "probability": { "type": "number", "minimum": 0, "maximum": 1 },
                        "status": { "type": "integer", "minimum": 100, "maximum": 599 },
                        "body": {},
                        "floorMs": { "type": "integer", "minimum": 0, "maximum": 60000 },
                        "jitterMs": { "type": "integer", "minimum": 0, "maximum": 60000 }
                      },
                      "required": ["kind"],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 50
                  }
                },
                "required": ["name", "rules"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "orgId": { "type": "string" },
                    "name": { "type": "string" },
                    "description": { "type": "string", "nullable": true },
                    "rules": { "type": "array", "items": {} },
                    "builtIn": { "type": "boolean" },
                    "createdAt": {
                      "anyOf": [{ "type": "string" }, { "type": "string", "format": "date-time" }]
                    }
                  },
                  "required": ["id", "orgId", "name", "rules"],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/v1/chaos-presets/{id}": {
      "delete": {
        "summary": "Delete a chaos preset",
        "tags": ["Chaos Presets"],
        "description": "Delete a chaos preset owned by the caller's org. Built-in presets and cross-org ids return 404.",
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:4000", "description": "local dev" }],
  "security": [{ "apiKey": [] }, { "bearerAuth": [] }, { "sessionCookie": [] }],
...
}